refactor backend models naming
