+++ blogging-platform/backend/routes/usersRoute.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import { User } from '../modals/usersModal.js';
+import { User } from '../model/usersModel.js';
 
 const router = express.Router();
 // route to add new user
